pyproject.toml
[project]
name = "mirascope-website-snippets"
version = "0.1.0"
description = "Python code snippets for Mirascope website"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Core dependencies
    "mirascope>=1.22.2",
    "pydantic>=2.7.4,<3.0",
    "pydub>=0.25.1,<1.0",
    "pillow>=10.4.0,<11.0",
]

[project.optional-dependencies]
# All provider dependencies
providers = [
    "openai>=1.6.0,<2",
    "anthropic>=0.29.0,<1.0",
    "mistralai>=1.0.0,<2",
    "cohere>=5.5.8,<6",
    "groq>=0.9.0,<1",
    "litellm>=1.42.12,<2",
    "azure-ai-inference>=1.0.0b4,<2.0",
    "aiohttp>=3.10.5,<4.0",
    "boto3>=1.34.70,<2",
    "boto3-stubs[bedrock-runtime]>=1.35.32,<2",
    "google-generativeai>=0.4.0,<1",
    "google-genai>=1.2.0,<2",
    "pillow>=10.4.0,<11",
]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.396",
    "ruff>=0.1.6",
]

[tool.pyright]
include = [
    "public/extracted-snippets",
    "public/examples"
]
exclude = [
    "**/__pycache__"
]